requirements.txt
Flask
gunicorn
google-generativeai>=0.5.0
python-dotenv
ffmpeg-python
pypdf
chromadb
langchain
requests